+++ RELEASES.md
@@ -42,6 +42,7 @@ Upcoming
 - Add fetch action for CmisSender in order to retreive metadata
 - Add CMIS bridge functionality (CmisListener) and the option to remap properties on getObject requests. A bridged cmisListener has to be configured to route traffic!
 - Replace struts page "Show configuration" by rest service
+- Add concurrency limit to parallel senders